word-project-2017-finel-project/requirements.txt
python-docx
colorama
